refactor(orb-live): A5 - extract tool catalog (VTID-02902)
Lifts buildLiveApiTools (~1,850 lines) out of orb-live.ts into a focused
module. Same function, same callers, same admin-tool injection. Zero
behavior change.

New file: services/gateway/src/orb/live/tools/live-tool-catalog.ts
  - export function buildLiveApiTools(mode, currentRoute, activeRole)
  - imports ADMIN_TOOL_SCHEMAS directly from '../../../services/admin-voice-tools'

orb-live.ts re-exports buildLiveApiTools so external callers (including
the A0.1 tool-catalog characterization test) continue to resolve from
this route file unchanged.

Verification:
  - 154 characterization tests pass, 9 snapshots match (the A0.1 tool-
    catalog 4-cell matrix snapshot proves the lifted catalog is
    byte-identical to the pre-A5 catalog).
  - tsc --noEmit: zero new errors.

orb-live.ts size impact:
  pre-A5:  15,727 lines
  post-A5: ~13,876 lines (−1,851)
  cumulative reduction from baseline: ~3,073 lines (−18%).

Part of Track A, step A5. A6 splits executeLiveApiToolInner() into
per-domain handlers next.
